add a preliminary un/pack tool
This adds the ability to use ffmpeg to pack grt files into mkv files together
with additionally media. One label track is stored as a subtitle. This is very
pre-liminary, more throught needs to go into the design of this
storage/recording option. Data is currently encoded as 32-bit float in the
wavpack lossless audio codec.